chore: `init` to depend on `add.Configure`
